requirements.txt
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0 